requirements.txt
discord.py
openpyxl
flask